refactor: 프론트엔드 코드를 src/web 디렉토리로 분리 - getWebPage 함수를 workers/app.ts에서 제거 - TemplateService 클래스 생성하여 HTML 템플릿 관리 - Dynamic import를 사용하여 템플릿 로드 - 메인 템플릿과 fallback 템플릿을 별도 파일로 분리 - 관심사 분리 원칙에 따라 프론트엔드 코드 구조화
